Add single-line comment syntax to CJSX [Fixes #40]
Before this change it was not possible to comment inside CJSX templates.

* https://github.com/jsdf/coffee-react-transform/issues/40
* https://github.com/jsdf/coffee-react/issues/13

After this change, you can use the syntax:

     {#this is a cjsx comment}

The syntax was designed to look like a CJSX_ESCAPE that contains a
comment, but as the resulting coffee-script would be illegal this is
special-cased by the parser to return nothing.

Future work would be needed make this syntax work inside a tag:

    <Person {#comment} />

or to make multiline comments work:

    <Person>{###

    ###}</Person>